Change refcap of JsonDoc to ref for better usability
JsonType has not been sendable anyways as JsonObject and JsonArray have been of refcap ref in JsonType,
so this only changes JsonDoc to be non-sendable, which should only have very little effect.

My first iteration was to make JsonDoc, JsonObject and JsonArray iso but this turned out to be cumbersome to use (especially when traversing into nested structures).
The other possibility would be to make all elements of JsonType val but this also turns out to be a little limiting when constructing a JSON structure, so i went with
the least invasive way for now.