Revert "Remove exception parameterisation from serialisation (#2524)"
This reverts commit d34dce667accf0797f0afcc5e04fd0605d1ad82e.

While performance testing the Pony version of Wallaroo's Market Spread
application in a multi worker scenario, it was discovered that the above
commit introduced a performance regression in both throughput and latency.
This regression is not present in a single worker run where serialisation
is not used. It should be noted that this regression only occurs when
consuming messages at a maximum rate and is not present if the sending
worker is not performing at maximum capacity. With this commit reverted,
sending in messages at the same rate resulted in higher throughput
and lower latencies.

If an example application is needed to recreate this behavior we can
provide instructions to run the Wallaroo application or attempt to
create a minimal example.