Fix incorrect tuple handling
I don't know the compiler well and I went off what I knew and believe I
found the root cause for our tuple bugs that started with 9222f08. Which
was in turn fixing https://github.com/ponylang/ponyc/issues/2735.

However, a number of bugs came from that fix because, it didn't fix the
root issue. Which was `gen_localdecl` returning GEN_NOVALUE where
GEN_NOTNEEDED was needed. This is why, prior to 9222f08, `gen_tuple` was
returning GEN_NOTNEEDED when the value it has was GEN_NOVALUE.

Fixes #2760