Fix GC-safety issue with writev pointers in File.
Prior to this commit, there was a garbage-collection safety issue
with the way that the `_pending_writev` array worked in the `File`
class. The pending writev data was being stored in a flat array
with element type of `USize`, with the byte pointers being converted
to type `USize` in order to be stored alongside the buffer length
numbers in the same array.

Storing these pointers as numbers meant that they are no longer
traced by the garbage as being pointers, meaning that the garbage
collector would try to free the underlying memory due to no longer
being pointed to by any valid pointer.

This PR fixes the issue by using an `Array[(Pointer[U8], USize)]`
instead of an `Array[USize]`. This also eliminates some complexity
in the code that was related to accessing the array data in blocks
of 2 elements per logical element.

This resolves part of https://github.com/ponylang/ponyc/issues/2526.
We could resolve the issue entirely by doing the same fix for
TCPConnection, though that may have some addiitonal complexity from
needing to be compatible with IOCP on windows.