Do not set File._errno when reading less than requested bytes.
When requesting more bytes than available from a File the _errno was set to
although reading less than available bytes (> 0) is no error condition.
If anything then FileEOF should be set in this case to signal that we reached the end of file.
Nonetheless i kept it as it is right now, as i think it is more reliable to detect FileEOF when reading 0 bytes.