Non-x86_64 CI via cross compiling and qemu
This commit adds CI builds for non-x86_64 targets using a combination
of cross compiling and qemu.

The architectures included are:

arm
armhf
aarch64
i686

This commit also includes some changes to ponyc and it's Makefile so
that cross compiling is a little easier to accomplish. Also included
are a couple of minor fixes related to getting CI to pass on arm
targets.

Lastly, this commit includes changes to the CI dockerfiles to make
the different ubuntu based docker images more consistent by having
them all inherit from a single base.