Deb packaging for bintray to allow for old version installation
With the introduction of deb packaging/distribution via Ubuntu
Launchpad PPA, we lost the ability for folks to be able to install
old versions of ponyc via their package manager. See:
https://github.com/ansible/ansible/issues/23143 for more info.
This commit switches back from Launchpad to Bintray for
hosting/distribution of deb packages. This necessitated that we
build the deb packages ourselves.

This PR includes changes to the travis ci config to now run two
additional jobs for every release build. One to build all the
debian version packages. The other to build all the ubuntu version
packages. The changes also include uploading the packages to a new
bintray debian repo.

Unfortunately, I ran into some issues with using `sbuild` on trusty
so I had to fall back to using docker containers for building the
packages for the different distros.