Add lib/llvm
This PR It supports both system installed LLVM binaries and vendored LLVM
sources. The default is the system installed LLVM binaries with the
vendored LLVM sources supported if make -f Makefile-lib-llvm is used.

The Makefile is a symlink to Makefile-ponyc so the system is built just
as it currently, no changes.

To build with the vendored LLVM, you may do "make -f Makefile-lib-llvm" or
change the Makefile symlink to point to Makefile-lib-llvm.

When using Makefile-lib-llvm lib/llvm is built first and then
Makefile-ponyc is invoked as Makefile-ponyc assumes llvm-config is
already installed.

The possible targets are those defined by Makefile-pony plus the following:

  clean         Clean only Makefile-ponyc does not touch lib/llvm/Makefile
  cleanall      Clean Makefile-ponyc and the build artifacts of
                lib/llvm/Makefile, does not touch lib/llvm/src which
                contains the source code for llvm
  distclean     Does a cleanall plus deletes lib/llvm/src
  some-tests    Runs ./build/debug/libponyc.tests with whatever is passed
                in the gtest_filter parameter.

These rebuild targets are used for development and will invoke the clean
target on Makefile-ponyc, then incrementally rebulid lib/llvm and then
build ponyc. Depending upon the

  rebuild       Rebuilds only does not run any tests
  rebuild-tests	Rebuilds and then invokes the "test" target on
                Makefile-ponyc
  rebuild-some-tests
                Rebuilds and then recursively invokes the some-tests
                target of Makefile-lib-llvm above. So whatever tests are
                specified in gtest_filter parameter will execute.

The .circleci/config.yml has been modified to tests the vendored LLVM on
various Docker images added in .ci-dockerfiles. Currently Alpine doesn't
work because the Alpine release requires changes to work with the musl
C library. What could happen is that Alpine patches could be applied to
a branch in ponylang/llvm and then be used by ponyc by adding a new
llvm_proj target.

Regarding whether to use submodule or include the source code directly,
I prefer this PR using submodule as opposed to PR #2663
(wip-lib-llvm) which hosts the source code directly. For the following
reasons:

 - Smaller ponyc repository.
 - All LLVM source code will be in a fork hosted at ponylang/llvm.
 - No changes in workflow is necessary to compile ponyc using there
   installed LLVM binaries.
 - For developers and regular users that wish to test an LLVM version
   not supported by there system this can now be easily done using
   make -f Makefile-lib-llvm llvm_proj=llvm_xxx. If they'd like they
   can create a symlink, ln -sf Makefile-lib-llvm Makefile, so the
   -f Makefile-lib-llvm is not needed.
 - Developers can modify lib/llvm/src and test using llvm_proj=llvm-current
 - Can more easily support multiple versions of LLVM without having the
   pony repository containing the source code LLVM directly.
 - The source code for LLVM is a fork of the LLVM that the resides in
   ponylang and would be managed as separate repo. This has the benefit
   that it will be easy to keep in sync with the upstream project. It will
   be easier to make any customization's such as the need to support
   patches for Alpine. And it will be easier to upstream changes since the
   fork can be kept in sync.
 - LLVM is a huge project compared to ponylang/ponyc project. There is
   almost 10x the lines of code as per cloc so the stats for ponyc will
   be artificially skewed. In addition, the binary size of the ponyc repo
   will significantly increase.

I would say the major downside to git submodule is that newer versions of
git are needed where as in PR #2663 "any" version of git could be used.
Specifically, in PR #2663 the git installed by apt-get in
ponylang/ponyc-ci:ubuntu is 1.8.2 and it worked fine, but I had to the
Dockerfile to install a newer version to work with this PR. I chose to
install 2.17.1 as that what was on my Arch Linux system, I suspect it
doesn't have to be that new, but it did work for me so that's what I
used.

# Conflicts:
#	.circleci/config.yml
#	Makefile