Use correct flags variable for epoll asio auto resub logic
https://github.com/ponylang/ponyc/pull/2781 accidentally used the
wrong `flags` variable. It used `flags` when it should have used
`ev->flags`.

This PR corrects the code to use the `ev->flags` variable. This PR
also removes the check for `ASIO_DISPOSABLE` from the boolean logic
because the enum value for `ASIO_DISPOSABLE` is `0` (see asio.h)
and the check will always be true.

This PR also adds some comments to that part of the code to clarify
what is occuring and why.