Fix linux build linking issue for non-native builds
Prior to this commit, ponyc would fail when linking after compiling
an application with an error that it couldn't find libponyrt if
ponyc was built with `make arch=XXXXX`.

This commit fixes the code to ensure when `make arch=XXXXX` is used
to build ponyc that ponyc will look in the appropriate directory
for libponyrt.