Ensure bintray deb travis jobs don't try and run PPA/COPR/etc
I missed one minor detail in the bintray deb packaging PR as I was
cleaning things up for the PR and that caused the bintray debs
travis jobs to fail when they tried to install LLVM as part of the
normal release steps.

This commit fixes that to ensure the normal release steps don't
run during the bintray deb building jobs.