Remove Launchpad PPA in favor of new Bintray DEB packages
This commit removes the recently added Launchpad PPA generation now
that the new Bintray DEB packages are successfully building during
releases. The README is updated to direct users to use the new
Bintray DEB packages instead of the Launchpad PPA packages. This
was done mainly because Launchpad PPA does not allow installing of
older versions once a new version is released.

This commit also includes the details regarding how to use the
AppImage package that is now cre0ated and uploaded to Bintray.