Fix bug related to searching within a Reader
Prior to this commit, if an unsuccessful search was performed,
the `_distance_of` function would cache the last node checked
for a value and future searches would resume from the next node
after the cached node. Unfortunately, this results in incorrect
behavior (an error) if someone were to search for a value that
doesn't exist in the buffer and then search for a value that
does exist in the buffer. This is because the search for the
existing value picks up where the search for the non-existing
value left off and reaches the end of the buffer and determines
that the value doesn't exist. However, if the search were to be
conducted starting from the current position in the buffer, then
the existing value would be correctly found by the `_distance_of`
function.

This commit fixes this by removing the caching and always having
`_distance_of` restart it's search from the current position in
the buffer. It also add a couple of test cases to explicitly
check that the searches are working correctly.