Fix invalid allocation bug in runtime that allows for segfaults
Thanks to @malte for identifying the root cause of #2013.

This PR fixes things so that if there's a request for a `size_t`
sized allocation, it actually tries to allocate it instead of
pretending that it allocated it while not actually allocating
anything which results in memory clobbering and segfaults.

resolves #2013