Add check to Path.clean() for root directory
Consider https://playground.ponylang.io/?gist=632ab1520993f82692e5c552a90f4418

This commit adds a check to the cleaned path's size, and prevents deleting a
separator character if it is the only character in the path. This fixes the bug
where Path.clean("/") returns ".", and adds a unit test to ensure it stays
fixed.