libponyc: resolve relative paths in use "path:..." statements
The use of `pony_realpath` (which calls the POSIX `realpath` underneath)
is convenient, but brings a change in behaviour: before, we have passed
just about anything into `prog->libpaths`, regardless of the directory's
existence. With `realpath`, nonexistent directories will make the call
return an error.

To remediate the impact a bit, I've added an `errorf()` statement.

However, to make this round, I think we might have to either check for
existence in both cases, or don't do that in either one.

Signed-off-by: Stephan Renatus <srenatus@chef.io>