Update the Windows Appveyor test script to compile most of the examples (#2993)
Update the Windows Appveyor test script to compile most of the examples.

This change adds a new target `examples` to the Windows build script.  For release builds, this will build all the programs under the `examples` directory, including the FFI examples.  For debug builds it will do nothing, as debug builds are too slow for AppVeyor.

Fixes #2990