Switch to a single atomic bitfield for mute/throttle bookkeeping
Prior to this commit, the updated muting/throttling logic relied
on three different atomic variables to keep track of all the state
needed for throttling and muting. This resulted in unnecessary
contention across different threads due to the multiple atomic
operations required.

This commit changes things to store the three pieces of information
(`is_muted`, `mute_map_count`, and `extra_throttled_msgs_sent`) in
a single `uint64_t` bitfield (for the current implementation and
constraints it is possible to fit everything into a single `uint32_t`
bitfield if desired). This adds a bit of complexity to the atomic
operations (we have to use compare and exchange now instead of
simpler fetch add/sub) but it ensure we only have to have a single
atomic operation for all the bookkeeping. This also results in
keeping track of `is_muted` status simpler because it removes a
race condition between the `is_muted` and the other bookkeeping
variables used to determine `batch_size` due to all of them now
being stored in a single variable.

This commit also moves the atomic fields in `pony_actor_t` to
separate cache line because they are accessed by other actors and
scheduler threads along with the other items accessed by other
actors.