Change default to disable pinning of scheduler threads to CPU cores
Prior to this commit, if a random user happened to be running
multiple pony applications that are CPU intensive, they would run
into unnecessary CPU contention due to CPU pinning (at least on
Linux) due to the first few cores being the ones used by all the
applications even if the system may have more cores available. This
was particularly troublesome because users may not even be aware
that the application they're running is a pony application.

This commit changes the default to not pin scheduler threads to CPU
cores. It also removes the `--ponynopin` CLI argument and replaces
it with the `--ponypin` CLI argument to leave the control for
pinning as an option for advanced users without requiring all users
of pony applications to understand pinning and how to disable it.

Resolves #3023