Change `--ponyminthreads` default to `0`
Thanks to @slfritchie's hard work in #2926 and #2561 to find and
fix the bugs in the dynamic scheduler scaling logic, scheduler
thread 0 now also suspends and resumes correctly without issues
(in my testing).

This commit changes the default for `--ponyminthreads` from `1` to
`0`. The last time @slfritchie ran into issues was in March and
left a comment with the commands he used at:
https://github.com/ponylang/ponyc/pull/2561#issuecomment-367900941

I ran the commands from @slfritchie's comment (with minor changes
to ensure `--ponyminthreads=0`) using Pony 0.26.0 for a few hours
without any issues.