Redo CircleCI setup
- Uses workflows to execute jobs only if preceeding jobs have passed.

For example, if release build of LLVM 7.0.1 fails, there's little point
to also building the debug version or the lib/llvm version.

- Removes most lib/llvm tests

We were testing every possible matrix just about. We have no evidence we
need that. This PR cuts it down to testing CentOS 7 (as we have no other
Centos 7 coverage due to lack of required LLVM packages there) and a
standard "llvm default" release and debug run. If folks find that
lib/llvm test coverage of this nature is allowing through problems, we
can expand out and add more tests.