Clean up llvm-config options
Does a few small things:

1- orders our search in terms of preference where we prefer newer LLVMs
to
older ones. Generic llvm-config is considered the least desirable.

2- only search in the users PATH. don't go poking into "magic"
directories
where some tool might have installed a version looking for llvm-config.
If we do that, then it makes it impossible for a user with more than 1
LLVM installation to pick the version to use based solely on their PATH.