Run FreeBSD CI jobs in sequence (#3055)
In all likelihood if one fails, both will fail. We have a maximum of 2 slots
for running FreeBSD jobs on Cirrus. With this change, they will be run in
sequence. First the release build tests, then the debug build tests.

This will allow for more than 1 PR to be doing FreeBSD tests at the same time.

Yes, it makes the total FreeBSD testing time longer for a given PR, however,
CircleCI and its many jobs that get run is the limiting factor for how long
it takes to fully test a PR, not Cirrus. In the end, the extra time of running
sequentially gets washed out for each individual PR.