Stop trying to be clever when finding user's LLVM installation
Clever is bad. Eventually it bites you in the ass. Ask The Narrator, he
knows.

Prior to this commit, we were trying to be clever and helpful and would
bend over backwards to try and find the user's LLVM installation
whereever their package manager of choice might put it. This was, in the
end, a losing battle. We were constantly adding more and more
variations.

To make it worse, we tried to have a hierarchy where we would look for
LLVM 7 first then 6 etc. We did this by looking something like:

llvm-config-7
llvm-config-6

The problem with this approach was that some installations, wouldn't use
the `-7` notation. For example, if you download a release directly from
releases.llvm.org, then your llvm-config will simply be `llvm-config`.

I discovered this when, after installing LLVM 7 from releases.llvm.org,
I continued to have LLVM 6 used because the search loooked something
like:

llvm-config-7
llvm-config-6
... more stuff ...
llvm-config

In the end, that's all just too much clever.

With this change, we'll use the first `llvm-config` we find in your path
IF you haven't set the LLVM_CONFIG environment variable which should be
considered the preferred way to select an LLVM when building ponyc.

Closes #3072