Improve "can't find LLVM" error message
There's no indication when doing testing of a setup whether you
correctly set the environment variable to *anything* or if you set it
but to something that isn't installed. This clarity might help someone
in the future as I'm looking at CI issue right now and wishing I had
this.

[skip ci]