Increase number of times we try fetching from FreeBSD repos
Some of our CI jobs for FreeBSD are timing out trying to install
packages using pkg.

There are two values we can change to control this:

FETCH_TIMEOUT and FETCH_RETRY. They default to 30 (for seconds) and 3
(for times). Each is set in /usr/local/etc/pkg.conf.

This commit doubles the number of retries from the default of 3 and
should hopefully cut down on our failures encountered while trying to
install dependencies.