Ensure methods reached through a union are detected as reachable.
Prior to this PR, the fact of whether methods of a union member were marked
as reachable when called via a union (or intersection) reference was
accidentally depending on the order in which things were marked
reachable in the compiler execution. When a type was marked
as a subtype of the union before the method call was reached,
everything was okay, but when that order was reversed
(possible through indirection of a trait), things were not okay -
the methods seen as reachable through the union weren't copied
into being reachable on the subtype of the union as they were
supposed to.

This PR fixes that by ensuring this copy happens. Prior to this PR,
the copy was only happening for traits and interface subtypes,
but now it happens for union and intersection subtypes too.

Resolves #3096.