Add XorOshiro128starStar and SplitMix64 as new PRNGs.
xoroshiro128** is a new xoroshiro variant with less bias in the low bits than xoroshiro128**
but also with a small performance impact.

Also use new parameters for xoroshiro128+ as described in the NOTE here:
http://xoshiro.di.unimi.it/xoroshiro128plus.c

splitmix64 is a PRNG which only needs 64 bits of state and is primarily used
to seed xoroshiro and xorshift generators from 64 bits only using the new 'from_64bits' constructors on those.

splitmix64 should only be used when 64 bit of state is a hard requirement, otherwise use xoroshiro128+ or xoroshiro128**.