Several Improvements to the random package (#3135)
* use faster fixed-point inversion for Random.int

instead of floating-point multiplication.
But only use it if the platform supports native 128 bit ops, as we need to multiply with U128, because we cannot (yet) usesmaller types when generating e.g. U32 with Random.int. Waiting for specialized generic functions here.

Also added some benchmarks to show the performance increase.

http://www.pcg-random.org/posts/bounded-rands.html states it is just as biased as fp-mult version and a quick local test, generating a histogram showed proper uniform distribution.

Also added 'int_unbiased
a slower but unbiased function for generating bounded random numbers.

* Add XorOshiro128starStar and SplitMix64 as new PRNGs.

xoroshiro128** is a new xoroshiro variant with less bias in the low bits than xoroshiro128**
but also with a small performance impact.

Also use new parameters for xoroshiro128+ as described in the NOTE here:
http://xoshiro.di.unimi.it/xoroshiro128plus.c

splitmix64 is a PRNG which only needs 64 bits of state and is primarily used
to seed xoroshiro and xorshift generators from 64 bits only using the new 'from_64bits' constructors on those.

splitmix64 should only be used when 64 bit of state is a hard requirement, otherwise use xoroshiro128+ or xoroshiro128**.

* add comments to tests about source of random numbers to test against

* rename from_64_bits constructors to from_u64

* [skip ci] added manual changelog entry