Wake up suspended scheduler threads if there is work for another one
Prior to this commit, suspended scheduler threads would not be woken
up even if there was more than one actor to be run in parallel.

This commit changes things so that if a scheduler thread has more
than one actor that has work to be done, it will wake up a suspended
scheduler thread to run the additional actor in parallel.

Resolves #3151.