Fix cycle detector issue with checking for blocked actors
As part of PR #2800, I broke the cycle detector due to incorrect
incremental logic for querying actors for if they're blocked. I
relied on the hashmap internal array/bucket index value instead
of keeping tracking of # of actors queried for if they're blocked
during the current function call.

This commit resolves the logic issue described above by keeping
a running count of actors queried during the current function
call.

This commit also implements @plprofetes suggestion to make the
threshold for # of actors to query for if they're blocked into
a dynamic value. Specifically, the threshold is now the larger
of the `CD_MAX_CHECK_BLOCKED` constant (which is 1000) or 10%
of the total number of actors currently known to exist by the
cycle detector.

Resolves #2975.