Makefile: make openssl_1.1.1 and openssl_1.1.x valid default_ssl versions
This introduces an array of valid versions, and only errors out if the
passed default_ssl version is not among them.

Unfortunately, the list of valid default_ssl version in the help output
does not use that variable.

Also switches the default_ssl debug output to "info" (was "error").

-----

In the stdlib code, this commit renames openssl_1.1.0 to openssl_1.1.x:

Since the treatment is the same, we might as well call it what it is.
Apparently it's confusing to set default_ssl to openssl_1.1.0 when you
only have 1.1.1. So, let's simplify things by calling it openssl_1.1.x.

-----

In Makefile-ponyc, we now treat openssl_1.1.{0,1} as openssl_1.1.x
accordingly.

Also removes the need for default_ssl_valid.

Note that the introduction of pony_default_ssl was necessary; resetting
default_ssl to some over value is _impossible_ when it was provided on
the command line, e.g. `make default_ssl=something`.

See https://www.gnu.org/software/make/manual/html_node/Overriding.html
for details:

> If you specify a value in this way, all ordinary assignments of the
> same variable in the makefile are ignored; we say they have been
> overridden by the command line argument.

Signed-off-by: Stephan Renatus <srenatus@chef.io>