Add ability to build with sanitizers and rework Makefile use processing
This commit changes the Makefile logic for when `use=` arguments
are provided to list out which `use=` options are being enabled
and also to throw an error if an unknown `use=` option is specified.

This commit also adds the ability to build with some common google
sanitizers for detecting potential issues with the runtime and
compiler.